requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-cors-headers==4.4.0
django-rest-knox==5.0.1
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
mysqlclient==2.2.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
